tests/requirements.txt
pytest==8.3.5
docker==7.1.0
zabbix==1.3.1
python-dotenv==1.1.0
psycopg2==2.9.10
